backend/requirements_new.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
chromadb==0.4.22
google-generativeai==0.3.2
sentence-transformers==2.3.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
